Rolling back to Android 23 and build tools 23
When we target Android N and the latest build tools we are also bringing in a compile dependency on Java 8.  We would prefer this to be a major version change and have rolled back that change.  Additionally org.json classes stopped compiling so I added a compile time dependency on the Android 4.1 runtime (our minimum)

Ensure you are running the latest Android versions by updating your Android SDK and running the maven android sdk deployer